refact(rate-nps): implement tests for nps and rate

+++ src/main/java/com/alura/chalenge/application/rates/NPS.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class NPS {
     private Course course;
-    private Integer score;
+    private float score;
     private String nps;
 }
